Refactor huak::pyproject and pyproject stdin
